+++ packages/opa/RELEASES.md
@@ -518,3 +518,11 @@ Based on:
 - Speakeasy CLI 1.533.0 (2.578.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v1.7.10] packages/opa
+
+## 2025-08-10 00:33:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.600.2 (2.677.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.8.0] packages/opa